Bugfix - Ensure client side timeouts are honored (#6920)
Fixing bug when timeout helper was not initialized correctly and hence client side timeouts were not behaving as expected.

Without the `true` argument, the timer doesn't start. For example in `RetryPolicy` which is used everywhere, the timer doesn't start till the first operation is completed. That results in unexpected behavior, and the operations being called for the second time though the timer has expired.