Bump Microsoft.Azure.Management.MixedReality version
No substantive changes aside from the version bump.

A small handful of users received version 0.9.0-preview of this package, but
with different contents, which can cause build failures depending on which
feeds are used to download packages and whether the package is cached or
not. A new version will allow these users to upgrade and avoid these
problems.