Update to .NET Core 3.0 GA toolset (#7719)
* Update to .NET Core 3.0 GA toolset

As part of updating a new nuget warning about package icon url's showed
up so disabled that warning as we want to stay with urls instead of
embedding icons in our packages. At least until we can discover a good
reason we need to embed them.

* Only publish artifacts if the build succeeded

If you publish artifacts on a failed build and then
re-run the build you will 100% always break the re-run
with the following error:

To help with re-running we are only going to upload on a successful build.