Add support for nested polymorphic objects in Firewall Policy
+ The autogenerated code for Firewall Policy is incorrect since the
PolymorphicJsonConverter used does not support nested polymorphism

We have FirewallPolicyRule with Derived Types: FirewallPolicyFilterRule
and FirewallPolicyNatRule.

These derived types have one/more FirewallPolicyRuleCondition(s), with
Derived Types: NetworkRuleCondition and ApplicationRuleCondition.

We need to customize the initialization of the NetworkManagementClient
to add a polymorphicJsonConverter that accepts the two class types and
discriminator fields.