[Event Hubs Client] Track Two: Fourth Preview (DateTimeOffset Parsing)
When runtime metrics were not enabled, `DateTime.MinValue` was being set by the
track one client for date-related items.  For some timezones, attempting to
translate this into the `DateTimeOffset` type used by the track two client was
triggering a validation error in the framework code.

These changes will fix translation such that the min value is translated to
a `null` offset, as it should be.