Treat request failure to IDMS as unavailability
While the timeout we include often means that we do not observe request
failures (since our timeout fires before the overall operation fails) I
was running into issues locally where `HttpRequestException`'s were
flying out of this method (because of the attempt to hit an unroutable
network).

In #7669 we changed the `HttpClientTransport` to wrap all exceptions in a
`RequestFailedException` so lets take a failing request to this endpoint
to mean that IDMS isn't available.