Adding default idle timeout for amqpConnection object (#7944)
During a network communication issue, when packets are not sent being sent from client to service (or vice versa) this parameter dictates how long should the connection be alive. On windows, the connection gets dropped automatically after sometime. However, in Linux, the connection stays in the `ESTABLISHED` state (using `netstat`) even after several minutes. 
From the customer view, it looks like `MessageSender.Send()` is called, but the operation just times out and throws `ServiceBusTimeoutException` instead of communication exception, since the connection never gets into faulted state.
With this change, we are forcing that 1 minute of idle time on the connection would force it to drop the connection and reestablish when required.