[Event Hubs Client] Track Two: Fifth Preview (Immutable LastEnqueuedEventProperties)
Refactoring to make the last enqueued event properties an atomic read and remove the
race condition from in-place updates.  In order to avoid unnecessary allocations, the transport
consumer will simply track the last event that it received and create the metadata on demand.

To better reflect this behavior, the existing property was changed to a method in an attempt to
discourage consumers from accessing it's properties directly.