[Event Hub Client] Track Two: Fifth Preview (Client Type Hierarchy)
Overhaul of the client type hierarchy to enable Event Hub producer and
consumer types to become stand-alone clients without the need to funnel
all activities through the Event Hub client.

These changes include creation of a connection to an Event Hub, which the
new client types can implicitly create or which advanced developers may
create and pass to clients to request that they share a single connection
to the Event Hub.

This set of changes does not include removal of partition affinity from the
consumer or producer, nor does it add the additional operations for consuming
events across all partitions.   Those will appear in a future set of changes.