Fix builds to work with package references (#8596)
We are hitting https://github.com/NuGet/Home/issues/6795 in a couple
projects after we switched to package references. Some builds did not
work correctly because there was a mix of package and project
reference in the resolution graph. Note this only repro's if
/p:SkipDevBuildNumber=true is set which caused the package and
the project to have the exact same version number but one coming
from source and the other from nuget.

To fix this I had to change Azure.Core.Test project to take a
package reference to keyvault.

Also needed to update the eventhubs checkpoint test project to
remove the project dependency on the eventhubs test project which
caused a mix-match of package/project references for eventhubs.