[Event Hubs Client] Track Two (ReadEventsFromPartition)
These changes are focused on beginning to remove partiton affinity from the
Event Hub Consumer Client by refactoring the iterator such that it allows a
partition and starting position to be passed when creating the iterator.

Each instance of the iterator is now indepdendent, using its own transport
consumer with dedicated AMQP link so there is no longer a need to coordinate
for a topic/subscription style approach of ensuring multiple callers see the
same stream of events.

Pending:
    - Extend partition context with the ability to read the last enqueued event information
    - Refactor the Receive method to a dedicated class and offer a CreateReceiver replacement
    - Remove the shared inner transport consumer
    - Remove the partition identifier and starting position from constructors