[Event Hubs Client] Track Two (Remove Producer Partition Affinity)
The focus of these changes is to remove the ability to create an Event Hub
Producer client that is bound to a specific partition.  Instead, the producer
should allow requesting a specific partition during send, in the same manner that
a partition hash key can be used.

Included in these changes is also removing the non-batch Send methods from the
public scope, as they will not be part of the initial API surface.

Finally, the BatchOptions type has been renamed to CreateBatchOptions.