Add hints on the shape of Uris for storage service (#8774)
For our constructors that take a Uri, provide a small note about what
the Uri is likely to look like (this is helpful if you are not using a
connection string directly, so you know what the Uri is likely to look
like, i.e. the suffix is likely to be {blob|queue}.core.windows.net).