[Event Hubs Client] Track Two (Consumer Options Refactor) (#8820)
The focus of these changes is to refactor the scope of the options used with
an Event Hub Consumer client, moving the owner level and tracking of last enqueued
event information to the reader level.   This allows for them to differ for each
active reader, rather than having the client bound to the specified values.

Also included are tweaks to test timings with the goal of improving test stability
and determinism after the large overhaul.