[Event Hubs Client] Track Two (Code Sharing Infrastructure)
The focus of these changes is to introduce an infrastructure for sharing internal
constructs needed across projects in the Event Hubs ecosystem in a clear way that
avoids reaching directly into a given project for a link.

The intent of this approach is to ensure that we can avoid unduly marking internals
visible outside of tightly associated test scenarios and to be sure that the scope
and impact of changing shared assets is apparent.