[Event Hubs Client] Track Two (Migrate Processor Package)
The focus of these changes is to migrate the `EventProcessorClient` from
the core Event Hubs client library into the new Event Processor client
library and allow it to take direct dependencies on Azure Blobs Storage.

In its present state, the types have been realigned for sharing to prevent
the need to grant internals access, and the unit tests are in a good state.
The live tests for the main Event Hubs library pass, while those for the
Event Processor library have been largely ignored.

Some areas to revisit:

- Event Processor diagnostics tests need to be updated; they should make use
  of a mocked consumer client rather than working against the transport layer.

- `EventProcessorClient::RunPartitionProcessing` has been tweaked for the case
  where the wait time elapses without an event; there needs to be additional
  testing and tuning for this case.

- The Processor Manager for the test infrastructure has been adjusted, but
  is not in a polished state and is likely to have errors.  It will need
  additional attention to ensure test validity.

- The Live Tests for the `EventProcessorClient` have been disabled due to
  instability; these will need to be patched up and validated.