[Event Hubs Client] Track Two (Samples and Documentation)
The focus of these changes is to revise the samples for the Event Processor
client, extend the set of samples with additional scenarios and flesh out the
README content for the samples in each package.

Also included are some small refinements to the Event Processor client that
were deferred from feedback for an earlier pull request, a fix for an empty
partition context when the maximum wait time elapses, and ignoring a
test that requires follow-up analysis.