Don't expose EventData fields as Nullable<T>'s
Because we share the EventData type between the consumer and producer
clients, we exposed some fields as Nullable<T>'s if they were assigned
by the Event Hubs service when processing an event (since, when creating
an EventData to send to the service as part of a batch, these properties
would not yet have actual values).

In practice, these properties will always be set when consuming events
and never be set when producing events, while having the Nullable<T>'s
type signiture correctly expresses these semantics, it makes the API
worse to program against in practice. In production scinerios it's
unlikely that you would ever inspect these properties anyway, and in
consumption cases they will alwasy be present so the nullablity doesn't
actually buy you much, but at the expense of `.Value` calls or other
ways of dealing with the (never actually there) `null`s.

So, we change Sequence, Offset and EnqueuedTime to no longer be
nullable. Instead, they are initialized with sentinal values
(`long.MinValue` for `Sequence` and `Offset` and
`default(DateTimeOffset)` for `EnqueuedTime`).  The choice of
`long.MinValue` instead of `0` is intentional, bases on user feedback
from the Track 1 SDKs that zero was a bad sentinal value for these
properties since it could be incorrectly interpreted as semantically
meaningful.