[Event Hubs Client] Track Two (AmqpMessageConverter Fix)
The focus of these changes is to fix a misused variable which, in a specific
exception case, could trigger a null reference error that would mask the
actual exception.

Also included is an enhancement for doc comments to make the ownership and
semantics of the ReadOnlyMemory<T> in EventData clear to those consuming it.