[Event Hubs Client] Track Two (Project Configuration) (#9082)
The focus of these changes is to revisit the configuration for the Event
Hubs projects and remove overrides that are no longer necessary or relevant,
adhering more closely to the central engineering system.

As part of these changes, the override to disable our custom code analysis
has been removed, enabling analysis.  The results were analyzed, and I've
created suppressions where they were not applicable to Event Hubs or were
producing false positives.