[Event Hubs Client] Track Two (Processor Event Members Tweak) (#9115)
The focus of these changes is to re-scope visibility for the
On[[ EVENT]] methods used for raising events within the processor.  Previously,
these followed the expected .NET event pattern of being protected/virtual.

After discission, it was decided that these should not be exposed, as we
do not have intended scenarios around extending the processor client.