[Event Hubs Client] Track Two (Fix Processor Resume from Checkpoint)
The focus of these changes is to correct an issue where the Event Processor
client would resume from a checkpoint using the event that the checkpoint
was based on, resulting in that event being processed multiple times.

The proccessor should assume the checkpoint is based on the event last processed
and should, therefore, resume from the next event available in the stream.