[Event Hubs Client] Track Two (Flatten Exception Hierarchy) (#9209)
The focus of these changes is to flatten the hierarchy of specialized exception
types and instead offer context around the failure reason in the form of an
enumeration of well-known cases.  It is expected that this will remove several
types that were not directly actionable for developers using the library, but
served to provide context.