[Event Hubs Client] Track Two (Event Position Offset Inclusivity) (#9220)
The focus of these changes is to reintroduce the ability to specify that
an offset-based `EventPosition` is inclusive or not.

The `EventProcessorClient` has been updated to use this approach rather than
incrementing the checkpoint.  Because incrementing an offset when the associated
event was the last in the  stream causes an error, there was no reliable means
resume from a checkpoint and start with the next event in the stream without
potentially triggering an error.