[Event Hubs Client] Track Two (Test Stability)
The focus of these changes is to improve test stability, especially for
Live tests in the nightly runs.  Timing has been adjusted to allow for
instability in ARM, which is notoriously cranky when resources are manipulated
on-the-fly and to ensure all tests are using timed cancellation to ensure that
they do not hang.