[EventHubs] Add notes on AAD Authentication
For users who would rather use AAD to connect to their Event Hubs, add a
small note that this library supports using Azure Identity to do so. In
addition, make a note about needing to assign specific roles to the AAD
principal used, and provide a link to the service's documentation on AAD
so users can learn more about what role they may want to use and how to
assign it to a principal.