[Event Hubs Client] Track Two (Release Documentation)
The focus of these changes is to revise and enhance the documentation in the
README files for each of the Event Hubs packages and to provide details for
the change log.  Versions have also been changed in the projects to match the
change log.

Adding the generated public API surface for analysis during CI.