Support `AZURE_AUTH_LOCATION` (#9312)
* Support `AZURE_AUTH_LOCATION`

This change introduces a new type, `AuthFileCredential` which can be
used to support authenticating with an "SDK Auth File", which the `az`
CLI can write.  The file itself is just a JSON document that has
information about the Tenent, Client ID and Client Secret. Under the
hood, we just create a ClientSecretCredential with the information from
the SDK Auth File.

We also add `AZURE_AUTH_LOCATION` to the list of environment variables
that the `EnvironmentCredential` considers.