[storage] Fix `CreateIfNotExists` documentation (#9385)
When these convenience methods were created, we just copied the
documentation from the `Create` versions. This included a comment that
the method failed if the thing being created already existed, which was
wrong.

We also did not note that in the case where the thing you are trying to
create already exists, these methods return null (instead of a regular
`Response<T>`), which is useful information to have.

This corrects both of these cases.