[Event Hubs Client] Track Two (Publishing Tweaks)
The focus of these changes is to tweak some of the behaviors around the
publishing of events to increase reliability during a corner case scenario
where a batch with single event at maximum size is published by multiple
producer clients with a high degree of concurrency over time.

This triggered the service to begin throttling, which caused some instability
in the producer's use of the AMQP link for publishing.

These changes introduce:

  - A more fine-grained processing of AMQP-triggered exceptions so that retry
    decisions can be made more intelligently.

  - An adjustment to the calculated retry delay when the service signals throttling,
    ensuring that the recommended minimum backoff will be respected.  This
    delay has some minor randomization to avoid the "thundering herd" problem
    when publishing is taking place concurrently.

  - A change to the approach used when building event batches to better guard
    against a race condition in disposal of AMQP messages.  Previously, the
    message used to measure the size when building the batch was cached and
    then copied for Send operations.  This caused some non-deterministic behavior
    when the AMQP link was under load and was performing its own buffering.

    The non-volitile event data is now cached with the batch and used to
    generate the AMQP message to be sent.