[Event Hubs Client] Track Two (Publish Resilliency)
The focus of these changes is to address a corner case in which a failure when
publishing events for the iterator to read would fail to publish any events that
had been received from the Event Hub but not published to the iterator channel.
Instead, a new receive request would be made upon recovery and a new set of events
would be used.

Publishing will now resume publishing with the most recently recieved batch of
events and not make another receive request until more events are needed.