[Event Hubs Client] Track Two (Publish Resilliency) (#9601)
The focus of these changes is to address a corner case in which a failure when 
publishing events for the iterator to read would fail to publish any events that 
had been received from the Event Hub but not published to the iterator channel. 
Instead, a new receive request would be made upon recovery and a new set of 
events would be used. This behavior would only manifest when the call to channel 
WriteAsync produced the failure; other exception and recovery paths were not impacted.

Publishing to the iterator channel will now resume with the most recently received 
batch of events and not make another receive request until that set is exhausted.