[Event Hubs Client] Project File Tweaks (#9721)
The focus of these changes is to reset the version for non-product supporting
libraries and pin them at 1.0.0, since there is little value in managing them
to keep aligned with the accompanying product.   Also updating the attribute
used for version tracking, as the prefix use was made obsolete a while back
and I forgot to update these.