[Event Hubs Client] Test Infrastructure Retry Tweak (Tracks One and Two) (#9738)
The focus of these changes is to continue the game of whack-a-mole to properly
detect and retry transient failures coming from the management plane when
dynamically creating resources for test runs.  In this episode, support
is added directly for the HttpRequestException, as peeking at the inner
exception did not match the existing clauses.