[Event Hubs Client] Mocked Async Event Source Workaround
The focus of these changes is to work around an issue that appears to be
related to the test runner's asynchronous-to-synchronous bridge where an
IAsyncEnumerable<T> is used in conjunction with a background task.

A specific error is triggered within the CLR that I've not been able to
pin down to an absolute root cause.  I have been able to reproduce using
NUnit, a background task, and a variety of IAsyncEnumerable implementations
outside of the Event Hubs ecosystem.

Ultimately, the exception is benign but does potentially interfere with
proper testing of the exception handler.  This workaround addresses the
specific scenario.