[Event Hubs Client] `Event Processor<TPartition>` Skeleton
The focus of these changes is to flesh out the public skeleton of the
`EventProcessor<TPartition>` to match the design proposal, and author the
public-facing doc comments to accompany them.

After these changes, the full set of types for the event processor should
exist and be in the correct project/namespace to allow for implementation of
the behaviors to move forward.