[eventhubs] Clarify MaximumWaitTime documentation
When using the `EventProcessorClient` with a `MaximumWaitTime`, if the
timeout is reached, the `ProcessEventAsync` event is raised with an
empty `ProcessEventArgs` structure. The documentation refered to this as
a "null event" which is a little confusing, since `ProcessEventArgs`
itself is not null. Instead, the `HasEvent` property returns `false`.
Try to clarify this is a bit and add some remarks and cross references.