[identity] Support AZURE_CLIENT_CERTIFICATE_PATH (#10196)
This change adds support for setting `AZURE_CLIENT_CERTIFICATE_PATH` to
the path to an PFX or PEM. The private key corresponding to the
certificate must be unencrypted.

As part of this change, we now allow constructing a
`ClientCertificateCredential` from a path to a file on disk, in either
PFX or PEM format.

Supporting PFX is pretty straightforward, since there's great support
in the .NET. For PEM files, things are harder. While newer versions of
.NET Core have the primitives needed to be able to decode a PEM file and
get an `X509Certificate2` file, earlier versions of .NET Core (and
versions of .NET Framework) have gaps in support.

There are two major steps for reading these PEM files:

1. Decoding the PEM file to extract both the certificate and the private
key.
2. Attaching the private key to the `X509Certificate2` object.

For (1), The hard part is decoding the PKCS#8 encoded Private Key. In
.NET Core 3.0+, there is built in support via
`RSA.ImportPkcs8PrivateKey`, and on platforms which don't support this,
we've included a lightweight decoder based on some code from the .NET
Team that is able to decode this data. While the reader does not support
the entirity of the PKCS#8 spec, in practice it is sufficent for
decoding PEM files generated by common tools, for example, the Azure CLI
as a part of `az ad sp create-for-rbac --create-cert`.

For (2), There is an API which is present in .NET 4.7.2+ and .NET Core
2.0+, which allows you to assign the `RSA` object obtained by (1) to a
`X509Certificate2`.  While we can't depend on this method directly,
because we target NET Standard 2.0, we can do a reflection invoke.

Taken together, this allows us to support PEM certificates on .NET Core
2.0+ and .NET Framework 4.7.2.

From an implementation point of view, the constructors that take a path
to a PEM file do not actually attempt to load the file when the object
is being constructed (since this will require I/O) so instead we defer
reading the file until the first time GetToken or GetTokenAsync is
called. In cases where there is a problem with decoding the file, we
ensure `CredentialUnavailableException` is raised (this ensures things
like `ChainedTokenCredential` and `AzureDefaultCredential` are able to
correctly ignore these credentials, if needed.

Contributes To: #8309