[EventHubs] Clean up EventPosition.Equals logic
Our logic for implementing `EventPosititon.Equals` had some extra code
that would try to use `object.ReferenceEquals` as a way to see if one
EventPosition was equal to another.

However, `EventPosition` is a struct, so it's not clear how this check
could ever really return true, since our calls to `ReferenceEquals` is
going to box both `EventPosition` arguments, so they will never be in
the same place in memory.