[Event Hubs Client] Event Processor<T> Startup/Shutdown
The focus of these changes is to implement the startup and shutdown functionality
for the event processor, along with enhancing the tests to cover the previously
implemented constructors.

At this time, partition processing and load balancing are not implemented;
these areas remain stubbed to allow for testing and will be part of a new
workstream.