[Event Hubs Client] Test Fixes
The focus of these changes is to fix a race condition that is believed to have
been the root cause of the non-deterministic behavior in CI runs.  Riding along
are enhancements to the string version of the `EventPosition` and `LastEnqueuedEventProperties`
types, which are treated as opaque values and offered little insight for failed
equality checks or the structural composition.