[Event Hubs Client] Event Processor<T> Partition Management
The focus of these changes is to implement the management for partition
processors, allowing the main load balancing loop to start and stop processing
for individual partitions in a fault-tolerant manner and with delegation to
the correct handlers.

Also included are some minor tweaks to synchronization of other tests to ensure
stability across environments.