[Event Hubs Client] Event Processor Client New Base Class
The focus of these changes is to overhaul the `EventProcessorClient` to
derive from `EventProcessor<T>`, allowing it to take advantage of the
centralized behavior and significantly reduce its area of responsibility.

A small set of tweaks was also made to the `EventProcessor<T>` implementation
to address gaps found when doing the conversion, primarily around ensuring
testability.

Orthogonal, but riding along, is a small tweak to the live test timings for
the `EventHubConsumer`.