[Event Hubs Client] Event Processor and Blob Checkpoint Store Test Cleanup
The focus of these changes is to address some deferred feedback provided for
the `EventProcessorClient` and `BlobsCheckpointStore` tests.

The maximum value for parallel execution of tests has also been lowered,
as Azure is under heavier load than normal and we're seeing an elevated
set of timeout failures during nightly runs.