[EventHubs] Handle errors not tied to a partition
`EventProcessor{TPartition}` passes a `null` partition object when
calling OnProcessingErrorAsync if the error is not tied to a specific
partition. We were not correctly handling this case. We should be
creating an `ProcessErrorEventArgs` with a `null` Partition ID in this
case, but instead we would crash.

This issue was introduced as part of the refactoring effort to make
`EventProcessorClient` inherit from `EventProcessor`.