[Event Hubs Client] Consumer Live Test Refactor
The focus of these changes is to overhaul the suite of live tests for the
`EventHubConsumer` to address areas of timing sensitivity and instability.
The refactor moves to a more deterministic approach and ensures that
safety cancellation is applied to all tests in order to ensure there are
no hangs.

Event generation has also been centralized into a shared helper, with removal
of the similar set of helper methods that had existed within the `EventProcessorClient`
live test suite.