Address an additional bug in the extension method.
When the size of the current segment was larger than the size of the local array we should have returned the array and get a larger array.
However, the check was looking at the entire size of the buffer and comparing that to the lenght of the current segment -- this would always be false.

Without the change, we would have thrown an exception when writing a sequence in which the current segment was larger than the previous one.

Added tests for both array-backed buffers as well as for NativeMemory backed ones. This required that the Azure.Core.Tests project builds with /unsafe option.