[Event Hubs Client] Event Batch Tweaks
The focus of these changes is to add thread safety to the core batch operations,
ensuring that TryAdd, Clear, and Dispose are safe to call concurrently.  Also
fixed behavior with Dispose where the "locked during send" semantics were not
properly respected.

An internal-facing Clear operation was added, allowing the batch state to be reset, in
anticipation of architect approval, at which point things will be public.