[Event Hubs Client] Additional Resiliency  For AMQP Link Creation (#11307)
Because of the transient nature of AMQP objects, there exists a race condition
where an attempt to open a link my happen concurrently to the existing connection
closing - either by cancellation or due to the service force-closing the connection
due to timeout or other condition.  Currently, this failure is detected and surfaces
to callers as an `InvalidOperationException` with an unclear message.

The focus of these changes is to provide resiliency in this scenario, wrapping the
service exception into a form that denotes that it is safe to retry and provides
messaging that is more clear for customers.

In addition, because the factory for AMQP links is created as part of class-level state,
it is not able to respect the cancellation token provided for individual operations.
This increases the possibility of observing the race condition between link creation and
closing of the connection when a cancellation token is signaled.  Additional guards have
been put in place to redundantly check the cancellation token before link creation to
provide a better chance to respeect the token.