[Event Hubs Client] Live Test Improvements
The focus of these changes is to unify the timeout value applied as a
safety measure to prevent hangs, hooking it into the test environment to
allow for configuration to differ between the local and CI environments.

A few minor tweaks were also made to attempt to stablize tests that have
been misbehaving during nightly runs.