[Event Hubs Client] Processor Release Readiness (v5.1.0)
The focus of these changes is to prepare the Azure.Messaging.EventHubs.Processor
project for release, updating the global Azure.Messaging.EventHubs reference
version and returning the processor to a package reference.

Orthogonal, but included, is shifting the new Azure.Core.TestFramework project
reference for the solutions into an "External" folder to denote that they're not
exclusively a part of the local solution.