Update how we configure out DevOpsFeed (#11917)
You cannot access "Build.Repository.Name" variable
at template expansion time so we never have a DevOpsFeedId
setup correctly with the current changes.

Instead we now have the DevOpsFeedId setup as a template parameter
that we override for nightly builds if we are not publishing from
the default public azure/azure-sdk-for-net repo. This will help us
not accidently publish something publicly that we don't want to.