Increasing Search test retry delays considerably for throttling
Creating indexes on a free account gets throttled and the ends of our test
passes are failing.  I'm increasing the timeouts a lot while we explore what
works for our tests.  The default exponential retry policy will randomly jitter
any individual retry between 80% and 120% of the following targets with these
settings:

```
  1:  0:01
  2:  0:02
  3:  0:04
  4:  0:08
  5:  0:16
  6:  0:32
  7:  1:04
  8:  2:08
  9:  4:16
 10:  8:32
Max: 10:00
```

We cap any single request at 10 minutes.  If that's not enough to get around
index throttling, we'll need to change the type of resources we're using.