[Event Hubs Client]
The focus of these changes is to provide a cleaner surface for mocking
diagnostic event sources, preventing the default constructor from running and
ensuring the mock does not emit messages to the named event source nor
apply the standard traits hooked by the Azure listeners.

Language requirement blurbs in the README files for Event Hubs and Service Bus
have been updated to remove confusing wording and provide better guidance on
the use of C# 8 in the libraries.