[Event Hubs Client] Load Balancer Options
The focus of these changes is to allow for additional options to control
load balancing within the event processor types set the stage for implementing
different strategies for load balancing.  Also included are surfacing options within
the consumer client to allow tuning the prefetch count and batch size, to allow
performance tuning.