[Event Hubs Client] Enhanced Connection String Validation
The focus of these changes is to enhance connection string validation for
all clients to ensure that issues are detected where reasonable messages
can be displayed.  Previously, validation was delegated to the EventHubConnection
which left a gap when the connection was not eagerly initialized, such as with
the event processor types.