[Event Hubs Client] Tweak for LoadBalancer.IsBalanced (#13423)
The focus of these changes is to adjust the way that the load balancer
determines whether or not it is currently in a balanced state.  The new
approach will consider the load balanced if the minimum share of partitions
is owned and no attempt was made to claim another partition during the current
cycle.