[Event Hubs Client] Blob Checkpoint Store: Assume Checkpoint Update
The focus of these changes is to modify the approach used when creating
a checkpoint to assume that the blob exists and should be updated, falling
back to creating the blob when that invariant does not hold.

This is a needed efficiency improvement due to the behavior of Blob
Storage when soft delete is enabled.  Each time a checkpoint was created,
a new blob was being uploaded, overwriting any existing blobs but potentially
causing a soft delete.  When listing blobs to read checkpoints, performance
will degrade as more soft deleted versions were tracked, lasting for ~2 weeks
until the deleted snapshots were garbage collected.